Only send calls when the InCallService is non-null
Bug Reports came in regarding the InCallController sending calls to a
null InCallService binding.  A null pointer exception was thrown and
which lead to a failure in adding the outgoing meetings.

To prevent the calls from failing, Telecom should verify the
InCallService binding is non-null before sending the new call over to
the InCallService.  In the event the InCallService is null, an amomaly
report is generated to further investigate when this situation arises
and the fequency.

Flag: com.android.server.telecom.flags.do_not_send_call_to_null_ics
Bug: 345473659
Test: unit
Change-Id: I5b335596a1918575472bafad11cfc3c6408ec1a8