Resolve call audio refactor test failures
There were test breakages found in trunk staging in BT and Telecom CTS
tests which are being resolved as part of this CL. Ensure that we don't
try to set the communication device for a null device and update the
system audio state when we receive UPDATE_SYSTEM_AUDIO_ROUTE in the
controller.

This CL also resolves an issue with audio routing to the base route when
playing the end tone (which may override a user's previously requested
route). For instance, when a wired headset is plugged in and the user
switches to speaker, we receive ACTIVE_FOCUS when playing the end tone
which causes audio to route back to wired headset. Audio incorrectly
plays on wired headset in this instance so we can opt to ignore falling
back on the base route for end tones to resolve this.

Bug: 349394420
Bug: 349060651
Flag: com.android.server.telecom.flags.use_refactored_audio_route_switching
Test: atest CtsTelecomTestCases, atest BluetoothInstrumentationTests
Change-Id: If9c977d9c32f4ae1436007b7132cc3d9d52be605