Resolve call audio refactor routing for auto
This CL provides a temporary fix for auto on cf to address routing to
devices of type bus (default). The legacy routing just assumes routing
is on speakerphone but the communication device will never be set. This
issue was surfaced as a result of a NPE where default routing was
assumed to be either earpiece or speaker but in the case of auto, the
routing is done on an audio type of bus. For now, we will resolve the
routing as unknown in order to resolve the NPE and provide a more robust
fix in order to support this routing.

Bug: 355676630
Test: atest CallAudioRouteControllerTest
Flag:
com.android.server.telecom.flags.use_refactored_audio_route_switching

Change-Id: I3a2e8d31da72492daa259d7d4398af8f27b58e6b