Fix issue where focus is retained after call.
The new audio routing refactor code had an issue where the focus was
not released at the end of a call.  The root cause was that
CallAudioRouteController was not calling
`mCallAudioManager.notifyAudioOperationsComplete()` when focus goes
inactive.

The rest of the changes you see are just logging cleanups to standardize
how Audio/BT operations are logged so that external API calls are easy
to identify, and to add method names where they were missing in the
log statements.

Flag: com.android.server.telecom.flags.use_refactored_audio_route_switching
Fixes: 356920567
Test: Manually testing audio focus; ensured media playing resumes after
a call.
Test: Added new tests to the Telecom CUJ suite to cover audio focus loss
and gain during managed, self-managed and transactional voip incoming
and outgoing calls.
Test: Updated unit tests for CallAudioRouteController to ensure that we
signal audio operations complete to CallAudioManager when focus is lost.

Change-Id: I4deba2ee8224fa0de5348bf49f615ad45c1af361