Update cant_call_due_to_no_supported_service wording.
This string was used in the past when SIP calling was not available due
to it being disabled.  However, this situation now more commonly happens
when a watch relies on bluetooth to place calls; if bluetooth is turned
off the same generic message will arise.  Given this is the more common
case, changing wording.

Flag: NONE string resources are not flaggable.
Test: Manual test.
Fixes: 341979014
Change-Id: Ic46c78d83d254b448263332a3b64282948765fd3