avoid concurrentModExeption by iterating over a list instead of set
A crash report came in showing that iterating over a set in the
CallEndpointController methods can cause a
ConcurrentModificationException.

In order to avoid this exception, the set can be converted to an
ArrayList. ArrayLists have safe iteration and can be iterated over
without worrying about ConcurrentModificationExceptions.

Flag: EXEMPT type change for iteration
Fixes: 359377211
Test: presubmits
Change-Id: Ie5314072403c23079157bf977b7892a646eef993