disconnect self-managed calls stuck in starting state
VoIP calls that do not move out of the NEW, CONNECTING,
DIALING, or RINGING state after the TIMEOUT  will now be
disconnected.  This prevents stuck calls in the system
which can have affects on UI (e.g. lock screen stays on).

This change was manually verified by creating calls in
the CONNECTING and RINGING states and ensuring they
are disconnected and destroyed if they are not moved
out of the state before the timeout.

Bug:  360298368
Flag: com.android.server.telecom.flags.disconnect_self_managed_stuck_startup_calls
Test: manual
Change-Id: I1aeb99bb5b81df600d9aa1bac253ca8ce65f4f0c