Resolve fallback audio routing for SCO disconnect
When SCO audio disconnects and there aren't any pending audio route
changes in the new audio path, ensure that audio is routed out of the
SCO audio path to an available route. On cf instances, the audio routing
was stuck on BT because normally the route is changed once BT stack
notifies Telecom that the connection state for the HFP device has
changed but this was not occurring on cf. Instead, ensure that when BT
stack notifies that SCO audio is disconnected that the audio routing is
updated in the UI to reflect the new audio state. This resolves a PTS
test failure found in cf instances.

Bug: 357023397
Test: atest TelecomUnitTests
Test: Manual test to ensure that audio routing is updated on SCO audio
disconnect.
Flag:
com.android.server.telecom.flags.use_refactored_audio_route_switching

Change-Id: I96c34e31f14921f22dd557dd11a3a09bdc2039a7