Remove references to CallRecordingTonePlayer.
This code is not used any more.  It was added to provide an audible tone
that the user could hear in a call to know a call was recorded when an
app tried to record using the mic during a call.  The audio policies on
the device no longer allow this.  Call recording via the dialer app as
is present now does not actually trigger this code.

Further, this was a per-carrier option and is not enabled for any carriers.
Removing this for modularization cleanup because CallRecordingTonePlayer
has unresolvable hidden API references.  Given this code is not used and
is not useful any more, removing.

Note; this flag is not yet in nextfood; it will get there for 25Q2.

Flag: com.android.server.telecom.flags.telecom_resolve_hidden_dependencies
Test: Ran Telecom unit tests and verified no regression.
Fixes: 308017600
Change-Id: I7a4b7f7f7362e4e47d2ea90ee962c2fa6d07c601