Ensure current user is used when setting the PhoneAccountSuggestionService.
This CL updates the relevant TelecomShellCommand code path to ensure the current primary user is being used when setting the PhoneAccountSuggestionService for Telecom CTS test purposes in PhoneAccountSuggestionHelper. In HSUM mode, this should be user 10 and in non HSUM mode this should be user 0. Without this change, PhoneAccountSuggestionServiceTest tests were failing in HSUM mode.

Test: atest PhoneAccountSuggestionServiceTest
Bug: 358587742
Flag: EXEMPT bugfix
Change-Id: I70ac06b08f0b94126c2f88adb11479785499e5d1