Resolve referencing for preferred audio routing
Minor fix to resolve referencing for obtaining the proper audio route
from the preferred audio routing strategy. Currently we are using the
AudioDeviceInfo type for referencing instead of the AudioRouteType. This
was causing issues when holding a call with the audio routed to speaker.
Once the call was unheld, the audio routed to earpiece because the audio
route was unable to be resolved for the audio type that was passed in.

Bug: 365155632
Test: Manual
Flag: com.android.server.telecom.flags.use_refactored_audio_route_switching
Change-Id: If9cd248bba2d21620f49eb21a7056ffac9818e9a