Resolve baseline call audio routing
This CL resolves the baseline audio routing in the refactored path.
There were recent changes to resolve the map referencing done when
computing the previced device strategy which was causing audio to stay
on speakerphone even when it was toggled off in Dialer. Instead, we
should just ignore using the preferred device strategy when we're
calculating the baseline route.

Bug: 366059141
Test: Manual to verify speaker can be toggled on/off
Test: Manual verification on WhatsApp
Flag:
com.android.server.telecom.flags.use_refactored_audio_route_switching

Change-Id: I9e4adf249bc745a92e42c4f8ae3f33cdd3305be1