Resolve call supported routes issue
When connecting a wired headset during a call, a reporter noted that the
audio wasn't going through the headset. Upon closer inspection, the call
supported routes weren't updated to take into account the newly added
route. As a result, the wired headset route wasn't found from the
available call supported routes so the routing was ignored. This CL
ensures that the call supported routes are updated when the available
routes are updated.

This wasn't caught by unit testing as the update only occurred when the
foreground call was updated (UPDATE_SYSTEM_AUDIO_ROUTE). Consequently,
the available routes were used instead of the call supported routes. Now
that the routes are properly updated for the call supported routes,
those are being used in the existing tests so no additional tests need
to be added.

Bug: 363885916
Test: Manual
Test: atest CallAudioRouteControllerTest
Flag:
com.android.server.telecom.flags.use_refactored_audio_route_switching

Change-Id: If71913181c7d632488726a6e2951825dfc5ce056